requirements.txt
google-generativeai
openai
torch
torchvision
torchaudio
transformers
sentencepiece
sacremoses